helper/requirements.txt
wget==3.2
PyYAML==5.1.2